Refactor attempts charts to remove duplicated code
